requirements.txt
google-genai>=0.5.0
pydantic>=2,<3
